fix: update routing
