+++ src/app/App.tsx
@@ -1,0 +1,7 @@
+import { Router } from '@/router/router';
+
+function App() {
+	return <Router />;
+}
+
+export default App;
