+++ src/components/ActivityBanner/ActivityBanner.tsx
@@ -4,7 +4,7 @@ import {
 	VOCAB_ACTIVE_ROUTE,
 	VOCAB_INACTIVE_ROUTE,
 	VOCAB_LEARNED_ROUTE,
-} from '@/utils/constants';
+} from '@/router/routes';
 import styles from './ActivityBanner.module.scss';
 import IconDumbbell from '@assets/icons/icon_dumbbell.svg?react';
 import { activityBanner } from '@/static/vocabulary';
