--- src/components/App/App.module.scss
@@ -1,1 +1,0 @@
-@import '@styles/mixins';
