--- src/components/App/App.tsx
@@ -1,11 +1,0 @@
-import AuthUserDictionary from '@pages/AuthUserDictionary';
-
-function App() {
-	return (
-		<>
-			<AuthUserDictionary />
-		</>
-	);
-}
-
-export default App;
