+++ src/components/LangAndStatusFilter/LangAndStatusFilter.tsx
@@ -5,7 +5,7 @@ import {
 	VOCAB_ACTIVE_ROUTE,
 	VOCAB_INACTIVE_ROUTE,
 	VOCAB_LEARNED_ROUTE,
-} from '@/utils/constants';
+} from '@/router/routes';
 import styles from './LangAndStatusFilter.module.scss';
 import iconLanguages from '@assets/icons/icon_languages.svg';
 import iconNotActive from '@assets/icons/icon_not_active.svg';
