--- src/components/layouts/LayoutWhithNavbar/LayoutWhithNavbar.module.scss
@@ -1,8 +1,0 @@
-@import '@styles/colors.scss';
-
-.content {
-	display: grid;
-	grid-template-columns: 10px auto;
-	grid-template-rows: auto auto;
-	background-color: $neutral-background-second-color;
-}
