--- src/components/pages/AuthUserDictionary/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AuthUserDictionary';
