--- src/components/pages/Vocabulary/Vocabulary.module.scss
@@ -1,25 +1,0 @@
-@import '@styles/colors.scss';
-@import '@styles/mixins.scss';
-
-.inner {
-	// padding-left: 140px;
-	// padding-right: 40px;
-	margin: 0 auto;
-	background-color: $neutral-background-second-color;
-
-	@include media($min-width: 1750px) {
-		max-width: 2190px;
-	}
-
-	@include media($max-width: 1750px, $min-width: 1350px) {
-		max-width: 1170px;
-	}
-
-	@include media($max-width: 1350px, $min-width: 830px) {
-		max-width: 1000px;
-	}
-
-	@include media($max-width: 830px) {
-		max-width: 500px;
-	}
-}
