+++ src/components/ui/ButtonIcon/ButtonIcon.tsx
@@ -1,4 +1,5 @@
 import cx from 'classnames';
+import * as DropdownMenu from '@radix-ui/react-dropdown-menu';
 import styles from './ButtonIcon.module.scss';
 
 type ColorTheme = 'light' | 'dark';
@@ -8,6 +9,7 @@ interface Props {
 	colorTheme?: ColorTheme;
 	className?: string;
 	title?: string;
+	trigger?: boolean;
 	type?: 'submit' | 'reset' | 'button';
 	disabled?: boolean;
 	active?: boolean;
@@ -19,11 +21,22 @@ export const ButtonIcon = ({
 	colorTheme = 'dark',
 	className,
 	title,
+	trigger,
 	type = 'button',
 	disabled = false,
 	onClick,
 }: Props) => {
-	return (
+	return trigger ? (
+		<DropdownMenu.Trigger
+			className={cx(styles.button, styles[`button_${colorTheme}`], className)}
+			onClick={onClick}
+			type={type}
+			disabled={disabled}
+			title={title}
+		>
+			{children}
+		</DropdownMenu.Trigger>
+	) : (
 		<button
 			className={cx(styles.button, styles[`button_${colorTheme}`], className)}
 			onClick={onClick}
