+++ src/layouts/LayoutWhithNavbar/LayoutWhithNavbar.tsx
@@ -2,21 +2,20 @@ import Navbar from '@components/Navbar';
 import Header from '@components/Header';
 import Footer from '@components/Footer';
 import styles from './LayoutWhithNavbar.module.scss';
+import { Outlet } from 'react-router';
 
-interface Props {
-	children: React.ReactNode;
-}
-
-function LayoutWhithNavbar({ children }: Props) {
+function LayoutWhithNavbar() {
 	return (
-		<div className={styles.layout}>
+		<>
 			<Header />
 			<div className={styles.content}>
 				<Navbar />
-				{children}
+				<main className={styles.inner}>
+					<Outlet />
+				</main>
 				<Footer />
 			</div>
-		</div>
+		</>
 	);
 }
