+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './styles/global.scss';
 import { BrowserRouter } from 'react-router-dom';
-import App from './components/App/App';
+import App from './app/App';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
 	<BrowserRouter>
