+++ src/pages/Vocabulary/Vocabulary.module.scss
