+++ src/pages/Vocabulary/index.ts
