+++ src/pages/WordProfile/WordProfile.tsx
