+++ src/pages/WordProfile/index.ts
