--- src/utils/constants.ts
@@ -1,15 +1,0 @@
-export const LANGS_ROUTE = '/languages';
-export const COLLECT_ROUTE = '/collections';
-export const FAVOR_ROUTE = '/favorite';
-export const EXERCISES_ROUTE = '/exercises';
-export const ABOUT_ROUTE = '/about-us';
-
-export const VOCAB_ROUTE = '/vocabulary';
-export const VOCAB_ACTIVE_ROUTE = '/vocabulary/active';
-export const VOCAB_INACTIVE_ROUTE = '/vocabulary/inactive';
-export const VOCAB_LEARNED_ROUTE = '/vocabulary/mastered';
-
-export const VOCAB_DANGER_ROUTE = '/vocabulary/danger';
-export const VOCAB_POTENTIAL_ROUTE = '/vocabulary/potential';
-export const VOCAB_TROHIES_ROUTE = '/vocabulary/trophies';
-export const VOCAB_WORD_ROUTE = '/vocabulary/word';
